+++ VectorFieldSDASimulatorPackage/VectorFieldSDASimulator/detector.py
